refactor: FSD Phase 6d — 管理・通知・ノート系の単一画面フックを pages の model へ移す (FRESTYLE-170)
Phase 6b/6c と同じ方針で、単一画面専用フック 7 個を pages/<slice>/model へ移す。

- useAdminDashboard → pages/admin-dashboard/model
- useAdminMembers → pages/admin-members/model
- useAuditLog → pages/admin-audit-log/model
- useCompanyApplications → pages/admin-company-applications/model
- useNotification → pages/notifications/model
- useNotes / useNoteKeyboardShortcuts → pages/notes/model

useNotes が使う sortOptions(ノートの並べ替え定数)は notes 専用なので
pages/notes/config へ移した。それに伴い NoteSortMenu(sortOptions と対の
notes 専用コンポーネント、NotesPage のみが使用)も pages/notes/ui へ移して
整合させた。これで components 直下から notes 関連が消える。

pages 間 import が無いことを確認済み(境界 lint 緑)。既存テストの assert は
変更していない。テスト件数 1184 で不変。
