+++ frontend/src/pages/notes/model/__tests__/useNoteKeyboardShortcuts.test.ts
