+++ frontend/src/pages/notes/model/useNoteKeyboardShortcuts.ts
